date: routing to include archive pages along with moving the 404 to a page
