+++ src/App.scss
@@ -1,14 +1,16 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@400;700&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@400;700&display=swap");
 
-html, body {
-  overflow: none;
+html,
+body {
+    overflow: none;
 }
 
 body {
-    font-family: 'Montserrat', sans-serif;
+    font-family: "Montserrat", sans-serif;
     color: white;
 }
 
-*:focus, button:focus {
-  outline: none !important;
+*:focus,
+button:focus {
+    outline: none !important;
 }
