+++ src/pages/404/styles.scss
